chore: deduplicate compatibility code
