+++ orebfuscator-compatibility/orebfuscator-compatibility-paper/pom.xml
@@ -20,6 +20,12 @@
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-compatibility-bukkit</artifactId>
+			<version>${revision}</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
 			<groupId>io.papermc.paper</groupId>
 			<artifactId>paper-api</artifactId>
 			<version>${dependency.paper.version}</version>
